gurudas: Fix timezone
Post date (when not specified explicitly in the body of the post) is
being set to one day in the future on github deployment. This behaviour
is not being observed during local testing.

Try changing the timezone in the hopes that github deployment will
display the correct date via `DEFAULT_DATE = 'fs'`.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>