gurudas: Specify date explicitly in second blog post
Omitting this results in 'mtime' setting my article's date one day ahead
in the future. This may be due to how I deploy my website on github and
the datetime configuration of the Github Pages machine that deploys it.

Since I have no time to dig deeper into this, specify date of post
explicitly for now.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>