gurudas: Re-remove DEFAULT_DATE 'fs'
Bring back the removal of DEFAULT_DATE. The problem with it in the
context of my particular Github Actions-based deployment workflow is
that the last modified date for all the blog posts and indexes is set to
the time of deployment, possibly because these files are all written
anew during the time of deployment (the repo is synced afresh as a first
step).

When this is removed, the `lastmod` dates in the sitemap for the blog
posts are the dates that are specified in their bodies (correct) and are
the date+time of deployment for the main `index.html` and the
`blog/index.html` pages which is what should be the case.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>