gurudas: blog: Fix navbar weird behaviour
The navbar disappears when scrolling down and comes back when scrolling
up - just on this one page and nowhere else on the blog.

I haven't as yet figured out why this is happening, but I have
root-caused it to this, so put in this hack for now and figure it out
later.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>