gurudas: Add heading shifter plugin
Thanks to Github Copilot, I now can set headings in my markdown content
to sane and manageable levels and have the code handle the generation
appropriately.

Previously, I had to set the exact level of headers I needed in the
final HTML output. This got cumbersome but I didn't have the bandwidth
to fix it. Now, I just have to use regular heading levels in Markdown
starting from `#` and `##` and all the way down without having to
wonder: "Was it 4 levels or 5 levels?"

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>